+++ trunk/onlinestore/src/main/java/com/onlinestore/service/impl/CategoryProductServiceImpl.java
@@ -2,7 +2,6 @@ package com.onlinestore.service.impl;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,29 +11,28 @@ import com.onlinestore.service.CategoryProductService;
 
 @Service("categoryProductService")
 @Transactional
-public class CategoryProductServiceImpl implements CategoryProductService{
+public class CategoryProductServiceImpl implements CategoryProductService {
 
-	
 	private CategoryProductDao cpDao;
-	
+
 	public void createCategoryProduct(CategoryProduct categoryProduct) {
-		cpDao.create(categoryProduct);		
+		cpDao.create(categoryProduct);
 	}
 
 	public List<CategoryProduct> getCategoryProducts() {
 		return cpDao.getAll();
 	}
 
-	public CategoryProduct getCategoryProduct(Long id) {
+	public CategoryProduct getCategoryProduct(Integer id) {
 		return cpDao.get(id);
 	}
 
 	public void updateCategoryProduct(CategoryProduct categoryProduct) {
-		cpDao.update(categoryProduct);		
+		cpDao.update(categoryProduct);
 	}
 
-	public void deleteCategoryProduct(Long id) {
-		cpDao.deleteById(id);		
+	public void deleteCategoryProduct(Integer id) {
+		cpDao.deleteById(id);
 	}
 
 }
